// littleberry-ng/src/app/models/book.model.ts
export interface Book {
  BookId: number;
  Title: string;
  Author: string;
  Prefix: string;
  BookNumber: number;
  CallNumber: string;
  Subject: string;
  SubjectId: number;
  Status: string;
  Url: string;
  Notes: string;
  Barcode: string;
  Description: string;
  Edition: string;
  ISBN: string;
  Media: string;
  PubDate: string;
  Publisher: string;
  PubLocation: string;
}